on second thought, move format_path helper into file.
On second thought move format_path helper back into file. The idea
of realpath() and format_path() returning different values is an
edge case and creates problems of it's own as the package
must attempt to load it when it is required.

For those reasons, moving it back into main class file.